sievedir.c: move lib/sievedir.[ch] to imap/sievedir.[ch]
