+++ timsieved/actions.c
@@ -63,10 +63,10 @@
 #include "xstrlcat.h"
 #include "xstrlcpy.h"
 #include "imap/global.h"
+#include "imap/sievedir.h"
 #include "imap/sync_log.h"
 #include "imap/tls.h"
 #include "imap/version.h"
-#include "sievedir.h"
 #include "sieve/sieve_interface.h"
 #include "timsieved/actions.h"
 #include "timsieved/codes.h"
